+++ NEWS.md
@@ -14,8 +14,7 @@
 * Enhanced `radab` variables `ADAPBLFL`, `ABLFL`, `ADABLFL`, `PARCAT1`, `RELTMU`
 * Add new variable `lgTMATRSK` to `adtte`.
 * Add `TNE` to `PARAMCD` and `Total Number of Exacerbations` to `PARAM` in `adtte`.
-* Add `Treatment induced ADA (Y/N)`, `Persistent ADA (Y/N)`, `Transient ADA (Y/N)`,
-  `Time to onset of ADA (Weeks)`, and `ADA Duration (Weeks)` to `PARAM` in `adab`.
+* Add new ADA and NAb `PARAM`s in `adab`.
 
 # random.cdisc.data 0.3.13