+++ NEWS.md
@@ -1,4 +1,4 @@
-# random.cdisc.data 0.3.13.9016
+# random.cdisc.data 0.3.13.9017
 
 ### New features
 * Added new random dataset generator: Anti-Drug Antibody Analysis Dataset (`radab`).